requirements.txt
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
da-vinci==0.3.0
Django==2.0
django-cleanup==5.1.0
django-rest-swagger==2.2.0
djangorestframework==3.9.2
easy-thumbnails==2.7
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
openapi-codec==1.3.2
Pillow==8.0.1
pytz==2020.4
PyYAML==5.3.1
requests==2.24.0
shortuuid==1.0.1
simplejson==3.17.2
uritemplate==3.0.1
urllib3==1.25.11
